requirements/project.txt
amqp==5.2.0
ansible==10.3.0
ansible-core==2.17.3
ansible-runner==2.4.0
asgiref==3.8.1
astroid==3.2.4
asttokens==2.4.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
black==24.8.0
boto3==1.34.156
botocore==1.34.156
celery==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.0
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.8
Django==5.1
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-htmlmin==0.11.0
django-impersonate==1.9.4
djangorestframework==3.15.2
djhtml==3.0.6
docutils==0.21.2
executing==2.0.1
html5lib==1.1
idna==3.7
ipython==8.26.0
isort==5.13.2
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
kombu==5.4.0
lockfile==0.12.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mccabe==0.7.0
mypy-extensions==1.0.0
mysql==0.0.3
mysqlclient==2.2.4
numpy==2.0.1
packaging==24.1
pandas==2.2.2
paramiko==3.4.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
phonenumbers==8.13.42
pip-autoremove==0.10.0
pipdeptree==2.23.1
platformdirs==4.2.2
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
PyGithub==2.3.0
Pygments==2.18.0
pygraphviz==1.11
PyJWT==2.9.0
pylint==3.2.6
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
PyNaCl==1.5.0
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
resolvelib==1.0.1
s3transfer==0.10.2
setuptools==72.1.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.1
stack-data==0.6.3
tomli==2.0.1
tomlkit==0.13.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
wrapt==1.16.0
